+++ src/main/java/racingcar/view/ResultView.java
@@ -1,4 +1,7 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.Car;
+import racingcar.domain.RacingCycle;
 
 import java.util.List;
 import java.util.stream.Collectors;
