+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,9 +1,10 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
