+++ src/test/java/racingcar/domain/UserInputTest.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.UserInput;
 
 import java.util.List;
