+++ src/test/java/racingcar/util/RandomGeneratorTest.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.util.RandomGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
